+++ app/build.gradle.kts
@@ -18,33 +18,11 @@ android {
 
     signingConfigs {
         create("release") {
-            val localProps = Properties().apply {
-                val localFile = rootProject.file("local.properties")
-                if (localFile.exists()) {
-                    load(localFile.inputStream())
-                }
-            }
-
-            val keystorePath = System.getenv("KEYSTORE_PATH")
-                ?: localProps.getProperty("KEYSTORE_PATH")
-                ?: throw GradleException("KEYSTORE_PATH is not set.")
-
-            val keystorePassword = System.getenv("KEYSTORE_PASSWORD")
-                ?: localProps.getProperty("KEYSTORE_PASSWORD")
-                ?: throw GradleException("KEYSTORE_PASSWORD is not set.")
-
-            val keyAliasValue = System.getenv("KEY_ALIAS")
-                ?: localProps.getProperty("KEY_ALIAS")
-                ?: throw GradleException("KEY_ALIAS is not set.")
-
-            val keyPasswordValue = System.getenv("KEY_PASSWORD")
-                ?: localProps.getProperty("KEY_PASSWORD")
-                ?: throw GradleException("KEY_PASSWORD is not set.")
-
-            storeFile = file(keystorePath)
-            storePassword = keystorePassword
-            keyAlias = keyAliasValue
-            keyPassword = keyPasswordValue
+            val signingConfig = project.getSigningConfig()
+            storeFile = file(signingConfig.keystorePath)
+            storePassword = signingConfig.keystorePassword
+            keyAlias = signingConfig.keyAlias
+            keyPassword = signingConfig.keyPassword
         }
     }
