+++ build.gradle.kts
@@ -4,8 +4,8 @@ plugins {
     alias(libs.plugins.kotlin.android) apply false
     alias(libs.plugins.kotlin.compose) apply false
     alias(libs.plugins.android.library) apply false
-    id("com.google.firebase.firebase-perf") version "1.4.2" apply false
-    id("com.google.firebase.crashlytics") version "3.0.4" apply false
+    alias(libs.plugins.firebase.perf) apply false
+    alias(libs.plugins.firebase.crashlytics) apply false
     alias(libs.plugins.jetbrains.kotlin.jvm) apply false
     alias(libs.plugins.google.firebase.appdistribution) apply false
     alias(libs.plugins.google.gms.google.services) apply false
