+++ gradle/libs.versions.toml
@@ -3,7 +3,7 @@ agp = "8.10.1"
 android-youtube-player = "12.1.2"
 datastorePreferences = "1.1.7"
 firebaseBom = "33.16.0"
-firebaseCrashlytics = "3.0.4"
+firebaseCrashlytics = "3.0.6"
 coil = "2.7.0"
 hiltAndroid = "2.57"
 hiltAndroidCompiler = "2.57"
@@ -56,6 +56,7 @@ hiltCommon = "1.2.0"
 datastoreCoreAndroid = "1.1.7"
 foundation = "1.8.3"
 accompanistPagerVersion = "0.36.0"
+firebasePerf = "1.4.2"
 
 [libraries]
 # AndroidX
@@ -171,6 +172,7 @@ kotlin-serialization = { id = "org.jetbrains.kotlin.plugin.serialization", versi
 kover = { id = "org.jetbrains.kotlinx.kover", version.ref = "kover" }
 hilt = { id = "com.google.dagger.hilt.android", version.ref = "hiltAndroid" }
 firebase-crashlytics = { id = "com.google.firebase.crashlytics", version.ref = "firebaseCrashlytics" }
+firebase-perf = { id = "com.google.firebase.firebase-perf", version.ref = "firebasePerf" }
 
 [bundles]
