+++ logger/src/main/java/com/parise_2/firebase/firebase/FireBaseCrashlyticsLogger.kt
@@ -2,9 +2,12 @@ package com.parise_2.firebase.firebase
 
 import com.google.firebase.crashlytics.FirebaseCrashlytics
 import com.parise_2.firebase.repo.Logger
+import com.parise_2.logger.BuildConfig
 
 class FireBaseCrashlyticsLogger : Logger {
     override fun logException(exception: Exception) {
-        FirebaseCrashlytics.getInstance().recordException(exception)
+        if (!BuildConfig.DEBUG) {
+            FirebaseCrashlytics.getInstance().recordException(exception)
+        }
     }
 }
